rename and refactor classes
