+++ python_modules/dagster/dagster_tests/definitions_tests/asset_reconciliation_tests/test_auto_observe.py
@@ -1,7 +1,7 @@
 from dagster import AssetKey, DagsterInstance, observable_source_asset
+from dagster._core.definitions.asset_daemon_cursor import AssetDaemonCursor
 from dagster._core.definitions.asset_graph import AssetGraph
 from dagster._core.definitions.asset_reconciliation_sensor import (
-    AssetReconciliationCursor,
     get_auto_observe_run_requests,
     reconcile,
 )
@@ -107,7 +107,7 @@ def test_reconcile():
         asset_graph=asset_graph,
         target_asset_keys=set(),
         instance=instance,
-        cursor=AssetReconciliationCursor.empty(),
+        cursor=AssetDaemonCursor.empty(),
         materialize_run_tags=None,
         observe_run_tags={"tag1": "tag_value"},
     )
